feat(server): add static files serving
