+++ pkg/hayloft/app.py
@@ -1,51 +1,51 @@
-from flask import request, jsonify, send_file, send_from_directory, Response
-from flask_cors import CORS
-from hayloft.schema import app, db, sse, Event, Session
+from gevent import monkey; monkey.patch_all()
+from gevent import sleep
+from bottle import app, static_file, request, response, GeventServer
+from hayloft.schema import db, Session, Event
+from hayloft.cors import EnableCors
+from hayloft.sse import sse
 from importlib.metadata import version
+from pathlib import Path
+from typing import Dict
 import argparse
 import time
 
-CORS(app)  # for development
+app = app()
+app.install(EnableCors()) 
+path = str(Path(__file__).parent.resolve()) 
 
-
-@app.route("/", methods=["GET"])
+@app.get("/")
 def index():
-    return send_file("public/index.html")
+    return static_file("index.html", root=f"{path}/public")
 
+@app.get("/assets/<file:path>")
+def serve_assets(file):
+    return static_file(file, root=f"{path}/public/assets")
 
-@app.route("/assets/<path:path>", methods=["GET"])
-def serve_assets(path):
-    return send_from_directory("public/assets", path)
-
-
-@app.route("/event", methods=["POST"])
+@app.post("/event")
 def create_event():
-    body = request.get_json()
-    title = body["title"]
-    message = body["message"]
-    type = body["type"]
+    body = request.json
+    session_name = body.get("session")
+    title = body.get("title")
+    message = body.get("message")
+    type = body.get("type")
     event: Event | None = None
-    sess: Session | None = None
+    new_session: Session | None = None
 
     try:
-        session = db.session.execute(
-            db.select(Session).filter_by(name=body["session"])
-        ).scalar_one()
-        event = Event(title=title, message=message, type=type, session_id=session.id)
-        db.session.add(event)
+        session = Session.select().where(Session.name == session_name).get()
+        event = Event.create(session=session, title=title, message=message, type=type)
     except:
         created_at = int(time.time() * 1000)
-        event = Event(title=title, message=message, type=type)
-        sess = Session(name=body["session"], created_at=created_at, events=[event])
-        db.session.add(sess)
-    db.session.commit()
-    msg: dict
-    if sess is not None:
+        new_session = Session.create(name=session_name, created_at=created_at)
+        event = Event.create(session=new_session, title=title, message=message, type=type)
+    msg: Dict
+    if new_session is not None:
         msg = {
             "session": {
-                "id": sess.id,
-                "name": sess.name,
-                "created_at": sess.created_at,
+                "id": new_session.id,
+                "name": new_session.name,
+                "created_at": new_session.created_at,
             },
             "event": None,
         }
@@ -56,54 +56,47 @@ def create_event():
                 "title": event.title,
                 "message": event.message,
                 "type": event.type,
-                "session_id": event.session_id,
+                "session_id": event.session.id,
             },
             "session": None,
         }
     sse.publish(msg, type="stream")
     return "OK"
 
-
-@app.route("/sessions", methods=["GET"])
+@app.get('/sessions')
 def get_sessions():
-    sessions = db.session.execute(db.select(Session)).scalars()
-    return jsonify(
-        [{"id": s.id, "name": s.name, "created_at": s.created_at} for s in sessions]
-    )
+    sessions = Session.select()
+    return {"sessions": [{"id": s.id, "name": s.name, "created_at": s.created_at} for s in sessions]} 
 
-
-@app.route("/sessions", methods=["DELETE"])
-def remove_sessions():
-    # todo it via cascade later
-    db.session.query(Session).delete()
-    db.session.query(Event).delete()
-    db.session.commit()
-    return ""
-
-
-@app.route("/sessions/<int:session_id>/events", methods=["GET"])
+@app.get("/sessions/<session_id:int>/events")
 def get_events(session_id):
-    events = db.session.execute(
-        db.select(Event).filter_by(session_id=session_id)
-    ).scalars()
-    return jsonify(
-        [
+    events = Event.select().where(Event.session_id == session_id)
+    return { 
+        "events": [
             {"id": e.id, "title": e.title, "message": e.message, "type": e.type}
             for e in events
         ]
-    )
+    }
 
-
-@app.route("/listen", methods=["GET"])
+@app.get("/listen")
 def listen():
-    def stream():
-        messages = sse.listen()
-        while True:
-            msg = messages.get()
-            yield msg
+    response.set_header("Content-Type", "text/event-stream")
+    response.set_header("Cache-Control", "no-cache")
+    yield 'retry: 100\n\n' 
+    messages = sse.listen()
+    
+    while True:
+        msg = messages.get()
+        yield msg
+        sleep(0.05)
 
-    return Response(stream(), mimetype="text/event-stream")
+def start():
+    db.connect()
+    db.create_tables([Session, Event])
 
+    print(f'\033[96mHayloft {version(__package__)} starting up, open in your browser http://localhost:7000\033[0m')
+    print("Hit Ctrl-C to quit.")
+    app.run(host='localhost', port=7000, server=GeventServer, quiet=True)
 
 def cli():
     parser = argparse.ArgumentParser(description="Hayloft - UI tool for LLM frameworks")
@@ -112,4 +105,7 @@ def cli():
     args = parser.parse_args()
 
     if args.command == "start":
-        app.run(host="localhost", port=7000, debug=False)
+        start()
+
+if __name__ == '__main__':
+    start()
