+++ pkg/hayloft/public/index.html
@@ -5,7 +5,7 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Hayloft</title>
-  <script type="module" crossorigin src="/assets/index-3c19358f.js"></script>
+  <script type="module" crossorigin src="/assets/index-0cc0a4d1.js"></script>
   <link rel="stylesheet" href="/assets/index-7630e259.css">
 </head>
