+++ pkg/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "hayloft"
-version = "0.1.9"
+version = "0.2.0a0"
 description = "UI tool for LLM frameworks"
 authors = []
 license = "MIT"
@@ -10,9 +10,6 @@ keywords = ["llm framework", "llm app tracking", "ui for llm app"]
 
 [tool.poetry.dependencies]
 python = "^3.8"
-flask = "^2.3.2"
-flask-cors = "^3.0.10"
-flask-sqlalchemy = "^3.0.4"
 nanoid = "^2.0.0"
 requests = "^2.31.0"
 gevent = "^22.10.2"
